cargo_auto_template_new_wasm::main_mod::lib_mod

Module hello_mod

All the real code is inside modules in separate files (program logic).

This doc-comments will be compiled into the docs.

Functions

format_hello_phrase

format the hello phrase

format_upper_hello_phrase

format the hello phrase with uppercase name if it is already uppercase, return error with thiserror

make_uppercase

return uppercase