Crate cargo_auto_template_new_wasm

cargo_auto_template_new_wasm

Template for a minimal wasm project for browser
version: 0.0.2 date: 2024-05-01 author: web_server_domain repository: GitHub

Hashtags: #maintained #ready-for-use #rustlang #automation #workflow
My projects on GitHub are more like a tutorial than a finished product: bestia-dev tutorials.
I recommend using the CRUSTDE - Containerized Rust Development Environment to write Rust projects on Linux, isolated from your system.

This template

Just like cargo new makes a soft and gentle introduction to Rust projects and development, I want to make the same for an in-browser WASM project with

cargo auto new_wasm project_name

Extremely simple, just the basic moving parts and use cases.
This simplest template does not have a PWA implementation or dedicated web server app.

Development details

Read the development details in a separate md file: DEVELOPMENT.md

Releases changelog

Read the releases changelog in a separate md file: RELEASES.md

TODO

And code happily ever after…

Open-source and free as a beer

My open-source projects are free as a beer (MIT license).
I just love programming.
But I need also to drink. If you find my projects and tutorials helpful, please buy me a beer by donating to my PayPal.
You know the price of a beer in your local bar ;-)
So I can drink a free beer for your health :-)
Na zdravje! Alla salute! Prost! Nazdravlje! 🍻

//bestia.dev
//github.com/bestia-dev
//bestiadev.substack.com
//youtube.com/@bestia-dev-tutorials

Modules

main_mod 🔒

wsm

src/web_sys_mod.rs helper functions for web_sys, window, document, dom, console, Trying to isolate/hide all javascript code and conversion here.

Enums

LibraryError

all possible library errors for thiserror

Functions

wasm_bindgen_start

To start the Wasm application, wasm_bindgen runs this functions